feat: Docker support

+++ sbin/wordcount_example_docker.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 set -eou pipefail
-#trap cleanUp INT
+trap stopCluster ERR INT
+
+BRIDGEFOUR_ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 function createData() {
   mkdir -p /tmp/data
@@ -9,46 +11,42 @@ function createData() {
   rm -rf /tmp/in/*
   mkdir -p /tmp/out
   rm -rf /tmp/out/*
-  wget https://www.gutenberg.org/cache/epub/2600/pg2600.txt -O /tmp/data/war_and_peace.txt
+  if [[ ! -f "/tmp/data/war_and_peace.txt" ]]; then
+    wget https://www.gutenberg.org/cache/epub/2600/pg2600.txt -O /tmp/data/war_and_peace.txt
+  fi
   # 66030 lines => 4 workers => ~16500 lines per worker
   split -l 16500 /tmp/data/war_and_peace.txt /tmp/data/war_and_peace.txt.
   mv /tmp/data/war_and_peace.txt.a* /tmp/in
-  #touch /tmp/in/file{0..2}.csv
 }
 
 function startCluster() {
-  WORKER1_PORT=$WORKER1_PORT WORKER2_PORT=$WORKER2_PORT sbt leader/run &
-  LEADER_PID=$!
-  echo "Started leader 1 [$LEADER_PID]"
-  WORKER_PORT=$WORKER1_PORT sbt worker/run &
-  WORKER_1_PID=$!
-  echo "Started worker 1 [$WORKER_1_PID]"
-  WORKER_PORT=$WORKER2_PORT sbt worker/run &
-  WORKER_2_PID=$!
-  echo "Started worker 2 [$WORKER_2_PID]"
+  cd "${BRIDGEFOUR_ROOT}"
+  docker-compose up -d
 }
 
+
+function stopCluster() {
+  echo "Cleaning up"
+  cd "${BRIDGEFOUR_ROOT}"
+  docker-compose down
+}
+
+
 function startJob() {
-  start=$(curl -sS --location 'http://localhost:5555/start' \
+  echo "Starting job on $LEADER_PORT"
+  start=$(curl -sS --location "http://localhost:$LEADER_PORT/start" \
     --header 'Content-Type: application/json' \
     --data '{
                         "name": "Example job",
                         "jobClass": {
                             "type": "DelayedWordCountJob"
                         },
-                        "input": "/tmp/in",
-                        "output": "/tmp/out",
+                        "input": "/data/in",
+                        "output": "/data/out",
                         "userSettings": {"timeout": "2"}
                     }')
   JOB_ID=$(echo "$start" | jq '.jobId')
   echo "${start}"
-}
-
-function cleanUp() {
-  echo "Cleaning up"
-  kill -9 $LEADER_PID
-  kill -9 $WORKER_1_PID
-  kill -9 $WORKER_2_PID
 }
 
 function printWorkerState() {
@@ -57,13 +55,17 @@ function printWorkerState() {
 }
 
 sleep=10
-WORKER1_PORT=5554
-WORKER2_PORT=5553
+LEADER_PORT=5550
+WORKER1_PORT=5551
+WORKER2_PORT=5552
 
 # Create data
 createData
 # Start one leader, 2 worker
 startCluster
+# Wait for it to come online
+echo "Waiting"
+sleep 5
 
 # Start a job
 startJob
@@ -71,7 +73,7 @@ echo "Started job $JOB_ID"
 
 # Observe status on the leader
 while true; do
-  status=$(curl -sS --location "http://localhost:5555/refresh/$JOB_ID")
+  status=$(curl -sS --location "http://localhost:$LEADER_PORT/refresh/$JOB_ID")
   echo "${status}"
   typ=$(echo "$status" | jq '.type')
   if [[ $typ != '"InProgress"' && $typ != '"Halted"' ]]; then
@@ -88,7 +90,7 @@ done
 
 # Get data
 echo "Checking results"
-curl -sS --location "http://localhost:5555/data/$JOB_ID" | jq 'to_entries | sort_by(-.value) | from_entries' | head -n 5
+curl -sS --location "http://localhost:$LEADER_PORT/data/$JOB_ID" | jq 'to_entries | sort_by(-.value) | from_entries' | head -n 5
 
 # Cleanup
-cleanUp
+stopCluster
